+++ apps/theme-builder-app-e2e/cypress.config.ts
@@ -6,12 +6,12 @@ export default defineConfig({
     ...nxE2EPreset(__filename, {
       cypressDir: 'src',
       webServerCommands: {
-        default: 'pnpm exec nx run theme-builder-app:serve',
-        production: 'pnpm exec nx run theme-builder-app:serve-static',
+        default: 'pnpm exec nx run theme-builder-app:serve --port=4300',
+        production: 'pnpm exec nx run theme-builder-app:serve-static --port=4300',
       },
-      ciWebServerCommand: 'pnpm exec nx run theme-builder-app:serve-static',
-      ciBaseUrl: 'http://localhost:4200',
+      ciWebServerCommand: 'pnpm exec nx run theme-builder-app:serve-static --port=4300',
+      ciBaseUrl: 'http://localhost:4300',
     }),
-    baseUrl: 'http://localhost:4200',
+    baseUrl: 'http://localhost:4300',
   },
 });